requirements.txt
APScheduler==3.6.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
insta-scrape==2.1.2
instaloader==4.9.6
instascrape==0.1
lxml==4.9.2
outcome==1.2.0
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.29.0
rsa==4.9
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
telegram==0.0.1
Telethon==1.28.5
tornado==6.3.1
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validate-email==1.3
wsproto==1.2.0
